stdio-common: Fix UB on stdio-common/vfprintf-process-arg.c
On i686 debug/tst-sprintf-fortify-unchecked triggers:

UBSAN: Undefined behaviour in vfprintf-process-arg.c:41:57 negation of 9223372036854775808 cannot be represented in type 'long long int'

+++ stdio-common/vfprintf-process-arg.c
@@ -38,7 +38,8 @@ LABEL (form_integer):
     {
       long long int signed_number = process_arg_long_long_int ();
       is_negative = signed_number < 0;
-      number.longlong = is_negative ? (- signed_number) : signed_number;
+      number.longlong = is_negative ? (- (unsigned long long) signed_number)
+				    : signed_number;
 
       goto LABEL (longlong_number);
     }
